BSM-74: Rename package to be a package and not a module name
