+++ features/bsm/vaadin-masterpage/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -5,7 +5,7 @@
     <reference id="transactionOperations" interface="org.springframework.transaction.support.TransactionOperations" availability="mandatory"/>
     <reference id="businessServiceManager" interface="org.opennms.netmgt.bsm.service.BusinessServiceManager" availability="mandatory"/>
 
-    <bean id="transactionAwareBeanProxyFactory" class="org.opennms.features.vaadin.components.core.TransactionAwareBeanProxyFactory">
+    <bean id="transactionAwareBeanProxyFactory" class="org.opennms.netmgt.vaadin.core.TransactionAwareBeanProxyFactory">
         <argument ref="transactionOperations"/>
     </bean>
     <bean id="bsmUi" class="org.opennms.netmgt.bsm.vaadin.masterpage.BusinessServiceMasterPageUI" scope="prototype">
@@ -16,7 +16,7 @@
         <argument ref="blueprintContainer"/>
         <argument value="bsmUi"/>
     </bean>
-    <bean id="bsmMasterPageUi" class="org.opennms.features.vaadin.components.core.AdminPageNavEntry">
+    <bean id="bsmMasterPageUi" class="org.opennms.netmgt.vaadin.core.AdminPageNavEntry">
         <property name="url" value="admin/bsm/masterpage.jsp" />
         <property name="name" value="Business Services Master Page" />
     </bean>
