+++ features/vaadin-components/core/src/main/java/org/opennms/netmgt/vaadin/core/TransactionAwareBeanProxyFactory.java
@@ -26,7 +26,7 @@
  * http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.vaadin.components.core;
+package org.opennms.netmgt.vaadin.core;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
