+++ features/vaadin-components/core/src/main/java/org/opennms/netmgt/vaadin/core/UIHelper.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.vaadin.components.core;
+package org.opennms.netmgt.vaadin.core;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
