+++ features/vaadin-components/core/src/test/java/org/opennms/netmgt/vaadin/core/TransactionAwareBeanProxyFactoryIT.java
@@ -26,7 +26,7 @@
  * http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.vaadin.components.core;
+package org.opennms.netmgt.vaadin.core;
 
 import org.junit.Assert;
 import org.junit.Before;
