+++ features/vaadin-jmxconfiggenerator/src/main/java/org/opennms/features/vaadin/jmxconfiggenerator/ui/HeaderPanel.java
@@ -38,7 +38,7 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.Panel;
 import com.vaadin.ui.PopupView;
 
-import org.opennms.features.vaadin.components.core.UIHelper;
+import org.opennms.netmgt.vaadin.core.UIHelper;
 import org.opennms.features.vaadin.jmxconfiggenerator.Config;
 
 /**
