+++ features/vaadin-jmxconfiggenerator/src/test/java/org/opennms/features/vaadin/jmxconfiggenerator/data/UiModelTest.java
@@ -3,7 +3,7 @@ package org.opennms.features.vaadin.jmxconfiggenerator.data;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.opennms.features.vaadin.components.core.UIHelper;
+import org.opennms.netmgt.vaadin.core.UIHelper;
 
 public class UiModelTest {
